Read backups from staging instead of integration
The idea here is to mirror from staging instead of integration, so that
we can reduce the frequency that the integration databases are restored.
See https://github.com/alphagov/govuk-helm-charts/pull/2719 for more
context.

We created the role in AWS which allows govuk-s3-mirror to read this
bucket in https://github.com/alphagov/govuk-aws/pull/1842

There are many many resources in this project which have "integration"
in their names, which could be confusing. I'll rename these in a
subsequent commit.