Fix pubsub topic
Terraform apply was giving the following error:

    │ Error: Error creating notification config for bucket govuk-s3-mirror_govuk-database-backups: googleapi: Error 400: Invalid Google Cloud Pub/Sub topic. It should look like '//pubsub.googleapis.com/projects/*/topics/*.', invalid
    │
    │   with google_storage_notification.govuk_database_backups-govuk_knowledge_graph_dev,
    │   on pubsub.tf line 59, in resource "google_storage_notification" "govuk_database_backups-govuk_knowledge_graph_dev":
    │   59: resource "google_storage_notification" "govuk_database_backups-govuk_knowledge_graph_dev" {

The "It should look like '//pubsub.googleapis.com/...'" suggests that
the API has got fussy about the format of these. I'm guessing the ones
that already exist aren't a problem because the API used to be more
permissive, but because I'm creating this one new I'm hitting the
stricter API.

I've already cheekily applied this with the new format, and it appears
to be fine. I need to move the other pubsubs over from integration
anyway, so I can fix those as and when.